requirements.txt
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.9
django-import-export==2.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
jdcal==1.4.1
MarkupPy==1.14
more-itertools==8.2.0
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
wcwidth==0.1.9
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
